RSC: Refactor: extract common code for env vars
