+++ packages/vite/src/rsc/rscBuildServer.ts
@@ -6,6 +6,7 @@ import { build as viteBuild } from 'vite'
 import { getWebSideDefaultBabelConfig } from '@redwoodjs/babel-config'
 import { getConfig, getPaths } from '@redwoodjs/project-config'
 
+import { getEnvVarDefinitions } from '../envVarDefinitions'
 import { onWarn } from '../lib/onWarn'
 
 /**
@@ -42,51 +43,7 @@ export async function rscBuildServer(
     envPrefix: 'REDWOOD_ENV_',
     publicDir: path.join(rwPaths.web.base, 'public'),
     envFile: false,
-    define: {
-      RWJS_ENV: {
-        // @NOTE we're avoiding process.env here, unlike webpack
-        RWJS_API_GRAPHQL_URL:
-          rwConfig.web.apiGraphQLUrl ?? rwConfig.web.apiUrl + '/graphql',
-        RWJS_API_URL: rwConfig.web.apiUrl,
-        __REDWOOD__APP_TITLE: rwConfig.web.title || path.basename(rwPaths.base),
-        RWJS_EXP_STREAMING_SSR: rwConfig.experimental?.streamingSsr?.enabled,
-        RWJS_EXP_RSC: rwConfig.experimental?.rsc?.enabled,
-      },
-      RWJS_DEBUG_ENV: {
-        RWJS_SRC_ROOT: rwPaths.web.src,
-        REDWOOD_ENV_EDITOR: JSON.stringify(process.env.REDWOOD_ENV_EDITOR),
-      },
-      // Vite can automatically expose environment variables, but we
-      // disable that in `buildFeServer.ts` by setting `envFile: false`
-      // because we want to use our own logic for loading .env,
-      // .env.defaults, etc
-      // The two object spreads below will expose all environment
-      // variables listed in redwood.toml and all environment variables
-      // prefixed with REDWOOD_ENV_
-      ...Object.fromEntries(
-        rwConfig.web.includeEnvironmentVariables.flatMap((envName) => [
-          // TODO (RSC): Figure out if/why we need to disable eslint here.
-          // Re-enable if possible
-          // eslint-disable-next-line
-          [`import.meta.env.${envName}`, JSON.stringify(process.env[envName])],
-          // TODO (RSC): Figure out if/why we need to disable eslint here
-          // Re-enable if possible
-          // eslint-disable-next-line
-          [`process.env.${envName}`, JSON.stringify(process.env[envName])],
-        ])
-      ),
-      ...Object.entries(process.env).reduce<Record<string, any>>(
-        (acc, [key, value]) => {
-          if (key.startsWith('REDWOOD_ENV_')) {
-            acc[`import.meta.env.${key}`] = JSON.stringify(value)
-            acc[`process.env.${key}`] = JSON.stringify(value)
-          }
-
-          return acc
-        },
-        {}
-      ),
-    },
+    define: getEnvVarDefinitions(),
     ssr: {
       // Externalize everything except packages with files that have
       // 'use client' in them (which are the files in `clientEntryFiles`)
